app/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
python-dotenv==1.0.0
google-cloud-firestore==2.14.0
PyJWT==2.8.0
bcrypt==4.1.2
gunicorn==21.2.0
